backend/requirements.txt
fastapi==0.115.13
uvicorn[standard]==0.24.0
openai==1.35.0
supabase==2.0.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.11.7
PyJWT==2.8.0
httpx>=0.24.0,<0.25.0
psutil==5.9.6
python-jose[cryptography]==3.3.0
stripe==10.12.0
